vp9: store segmentation feature data as i16 (ALT_Q range ±255)
read_bits(8) as i8 silently wrapped magnitudes 128-255 for signed
segmentation features (ALT_Q, ALT_Y_AC are 8-bit magnitude + sign bit).
Corrupted values were fed to the Vulkan FeatureData and VAAPI qindex/LF
math. Store i16 in Vp9Segmentation::feature_data; the Vulkan std struct
already used i16.

+++ crates/vacc-core/src/picture.rs
@@ -1379,7 +1379,8 @@ pub struct Vp9Segmentation {
     /// Feature enabled flags [VP9_MAX_SEGMENTS]
     pub feature_enabled: [u8; VP9_MAX_SEGMENTS as usize],
     /// Feature data [VP9_MAX_SEGMENTS][VP9_SEG_LVL_MAX]
-    pub feature_data: [[i8; VP9_SEG_LVL_MAX as usize]; VP9_MAX_SEGMENTS as usize],
+    /// i16: ALT_Q/ALT_Y_AC are 8-bit magnitude + sign (range ±255), does not fit i8.
+    pub feature_data: [[i16; VP9_SEG_LVL_MAX as usize]; VP9_MAX_SEGMENTS as usize],
 }
 
 impl Default for Vp9Segmentation {
